moved parameters to objects.device folder
