+++ src/fdtdx/core/wavelength.py
