+++ src/fdtdx/objects/device/parameters/binary_transform.py
