+++ src/fdtdx/objects/device/parameters/discretization.py
