+++ src/fdtdx/objects/device/parameters/latent.py
